+++ Source/Game/Application/Configuration/Enums/PerformanceOverlayPreset.cs
@@ -22,7 +22,7 @@ terms, you may contact me via email at nyvantil@gmail.com.
 */
 
 namespace Game.Application.Configuration.Enums {
-	public enum PerformanceOverlayPreset : byte {
+	public enum PerformanceOverlayPreset : uint {
 		Hidden,
 		FpsOnly,
 		Partial,
