+++ Source/Game/Application/Configuration/Enums/ShadowFilterQuality.cs
@@ -22,7 +22,7 @@ terms, you may contact me via email at nyvantil@gmail.com.
 */
 
 namespace Game.Application.Configuration.Enums {
-	public enum ShadowFilterQuality : byte {
+	public enum ShadowFilterQuality : uint {
 		Off,
 		Low,
 		High,
